+++ previous-versions/sync-for-commerce-version-1/src/sdk/accountingcustomers.ts
@@ -87,7 +87,7 @@ export class AccountingCustomers {
         headers["user-agent"] = this.sdkConfiguration.userAgent;
 
         const globalRetryConfig = this.sdkConfiguration.retryConfig;
-        let retryConfig: any = retries;
+        let retryConfig: utils.RetryConfig | undefined = retries;
         if (!retryConfig) {
             if (globalRetryConfig) {
                 retryConfig = globalRetryConfig;